+++ examples/GetTokenomicsInfo/Program.cs
@@ -10,7 +10,7 @@ namespace GetTokenomicsInfo;
 internal sealed class GetTokenomicsInfoOptions
 {
     [Option(HelpText = "URL representing the endpoint where the gRPC V2 API is served.",
-        Default = "http://node.testnet.concordium.com:20000/")]
+        Default = "http://grpc.testnet.concordium.com:20000/")]
     public string Endpoint { get; set; }
 
     [Option(