+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Unreleased changes
+
 - Bugfix
   - Switched the GitHub runners from using 'ubuntu-latest' to 'ubuntu-20.04' to ensure compatibility with the default .NET 6 Docker image for the SDK.
 - Added
   - New GRPC-endpoint: `GetBlockItems`
-  - New transaction `DeployModule`
+  - New transactions: `DeployModule` and `UpdateContract`.
   - The function `Prepare` has been removed from the `AccountTransactionPayload` class, but is implemented for all subclasses except `RawPayload`.
   - Added serialization and deserialization for all instances of `AccountTransactionPayload`
 - Breaking